requirements.txt
django~=3.2
django-environ
django-filter~=22.1
django-health-check~=3.16
djangorestframework~=3.13
drf-yasg~=1.20
graphene-django~=2.15
gunicorn~=20.1
psycopg2-binary~=2.9
whitenoise~=6.2
